+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Added `knitr` engine `rtex` (see vignette "Chunk engine for easy tex output in Rmarkdown and Quarto")
 * New functions: `as_expr.caracas_solve_sys_sol`
+* `inv()` by default uses `gauss` like SymPy
 
 # caracas 2.0.1